+++ lib/Models/Cesium.ts
@@ -817,10 +817,9 @@ export default class Cesium extends GlobeOrMap {
     const vectorFeatures = this.pickVectorFeatures(screenPosition);
 
     const providerCoords = this._attachProviderCoordHooks();
-    const pickRasterPromise = this.scene.imageryLayers.pickImageryLayerFeatures(
-      pickRay,
-      this.scene
-    );
+    var pickRasterPromise = this.terria.allowFeatureInfoRequests
+    ? this.scene.imageryLayers.pickImageryLayerFeatures(pickRay, this.scene)
+    : undefined;
 
     const result = this._buildPickedFeatures(
       providerCoords,